fixed and added for tests for codon optimization and sequence structs.
