+++ io/genbank/genbank.go
@@ -107,18 +107,8 @@ func (sequence *Genbank) AddFeature(feature *Feature) error {
 	return nil
 }
 
-// GetFeatures returns a slice of features from a Genbank struct. Is equivalent to sequence.Features but for interfaces.
-func (sequence *Genbank) GetFeatures() ([]Feature, error) {
-	return sequence.Features, nil
-}
-
 func (feature Feature) GetSequence() (string, error) {
 	return getFeatureSequence(feature, feature.Location)
-}
-
-// GetType takes a feature and returns a sequence type for that feature.
-func (feature Feature) GetType() (string, error) {
-	return feature.Type, nil
 }
 
 // getFeatureSequence takes a feature and location object and returns a sequence string.
