--- io/poly/poly_test.go
@@ -1,120 +1,0 @@
-package poly_test
-
-import (
-	"fmt"
-	"testing"
-
-	"github.com/TimothyStiles/poly/io/genbank"
-	"github.com/TimothyStiles/poly/transform"
-)
-
-func ExampleSequence_AddFeature() {
-
-	// Sequence for greenflourescent protein (GFP) that we're using as test data for this example.
-	gfpSequence := "ATGGCTAGCAAAGGAGAAGAACTTTTCACTGGAGTTGTCCCAATTCTTGTTGAATTAGATGGTGATGTTAATGGGCACAAATTTTCTGTCAGTGGAGAGGGTGAAGGTGATGCTACATACGGAAAGCTTACCCTTAAATTTATTTGCACTACTGGAAAACTACCTGTTCCATGGCCAACACTTGTCACTACTTTCTCTTATGGTGTTCAATGCTTTTCCCGTTATCCGGATCATATGAAACGGCATGACTTTTTCAAGAGTGCCATGCCCGAAGGTTATGTACAGGAACGCACTATATCTTTCAAAGATGACGGGAACTACAAGACGCGTGCTGAAGTCAAGTTTGAAGGTGATACCCTTGTTAATCGTATCGAGTTAAAAGGTATTGATTTTAAAGAAGATGGAAACATTCTCGGACACAAACTCGAGTACAACTATAACTCACACAATGTATACATCACGGCAGACAAACAAAAGAATGGAATCAAAGCTAACTTCAAAATTCGCCACAACATTGAAGATGGATCCGTTCAACTAGCAGACCATTATCAACAAAATACTCCAATTGGCGATGGCCCTGTCCTTTTACCAGACAACCATTACCTGTCGACACAATCTGCCCTTTCGAAAGATCCCAACGAAAAGCGTGACCACATGGTCCTTCTTGAGTTTGTAACTGCTGCTGGGATTACACATGGCATGGATGAGCTCTACAAATAA"
-
-	// initialize sequence and feature structs.
-	var sequence genbank.Genbank
-	var feature genbank.Feature
-
-	// set the initialized sequence struct's sequence.
-	sequence.Sequence = gfpSequence
-
-	// Set the initialized feature name and sequence location.
-	feature.Description = "Green Flourescent Protein"
-	feature.Location = genbank.Location{}
-	feature.Location.Start = 0
-	feature.Location.End = len(sequence.Sequence)
-
-	// Add the GFP feature to the sequence struct.
-	sequence.AddFeature(&feature)
-
-	// get the GFP feature sequence string from the sequence struct.
-	featureSequence, _ := feature.GetSequence()
-
-	// check to see if the feature was inserted properly into the sequence.
-	fmt.Println(gfpSequence == featureSequence)
-
-	// Output: true
-}
-
-func ExampleFeature_GetSequence() {
-
-	// Sequence for greenflourescent protein (GFP) that we're using as test data for this example.
-	gfpSequence := "ATGGCTAGCAAAGGAGAAGAACTTTTCACTGGAGTTGTCCCAATTCTTGTTGAATTAGATGGTGATGTTAATGGGCACAAATTTTCTGTCAGTGGAGAGGGTGAAGGTGATGCTACATACGGAAAGCTTACCCTTAAATTTATTTGCACTACTGGAAAACTACCTGTTCCATGGCCAACACTTGTCACTACTTTCTCTTATGGTGTTCAATGCTTTTCCCGTTATCCGGATCATATGAAACGGCATGACTTTTTCAAGAGTGCCATGCCCGAAGGTTATGTACAGGAACGCACTATATCTTTCAAAGATGACGGGAACTACAAGACGCGTGCTGAAGTCAAGTTTGAAGGTGATACCCTTGTTAATCGTATCGAGTTAAAAGGTATTGATTTTAAAGAAGATGGAAACATTCTCGGACACAAACTCGAGTACAACTATAACTCACACAATGTATACATCACGGCAGACAAACAAAAGAATGGAATCAAAGCTAACTTCAAAATTCGCCACAACATTGAAGATGGATCCGTTCAACTAGCAGACCATTATCAACAAAATACTCCAATTGGCGATGGCCCTGTCCTTTTACCAGACAACCATTACCTGTCGACACAATCTGCCCTTTCGAAAGATCCCAACGAAAAGCGTGACCACATGGTCCTTCTTGAGTTTGTAACTGCTGCTGGGATTACACATGGCATGGATGAGCTCTACAAATAA"
-
-	// initialize sequence and feature structs.
-	var sequence genbank.Genbank
-	var feature genbank.Feature
-
-	// set the initialized sequence struct's sequence.
-	sequence.Sequence = gfpSequence
-
-	// Set the initialized feature name and sequence location.
-	feature.Description = "Green Flourescent Protein"
-	feature.Location.Start = 0
-	feature.Location.End = len(sequence.Sequence)
-
-	// Add the GFP feature to the sequence struct.
-	sequence.AddFeature(&feature)
-
-	// get the GFP feature sequence string from the sequence struct.
-	featureSequence, _ := feature.GetSequence()
-
-	// check to see if the feature was inserted properly into the sequence.
-	fmt.Println(gfpSequence == featureSequence)
-
-	// Output: true
-
-}
-
-func TestFeature_GetSequence(t *testing.T) {
-	// This test is a little too complex and contrived for an example function.
-	// Essentially, it's testing GetSequence()'s ability to parse and retrieve sequences from complex location structures.
-	// This was originally covered in the old package system  it was not covered in the new package system so I decided to include it here.
-
-	// Sequence for greenflourescent protein (GFP) that we're using as test data for this example.
-	gfpSequence := "ATGGCTAGCAAAGGAGAAGAACTTTTCACTGGAGTTGTCCCAATTCTTGTTGAATTAGATGGTGATGTTAATGGGCACAAATTTTCTGTCAGTGGAGAGGGTGAAGGTGATGCTACATACGGAAAGCTTACCCTTAAATTTATTTGCACTACTGGAAAACTACCTGTTCCATGGCCAACACTTGTCACTACTTTCTCTTATGGTGTTCAATGCTTTTCCCGTTATCCGGATCATATGAAACGGCATGACTTTTTCAAGAGTGCCATGCCCGAAGGTTATGTACAGGAACGCACTATATCTTTCAAAGATGACGGGAACTACAAGACGCGTGCTGAAGTCAAGTTTGAAGGTGATACCCTTGTTAATCGTATCGAGTTAAAAGGTATTGATTTTAAAGAAGATGGAAACATTCTCGGACACAAACTCGAGTACAACTATAACTCACACAATGTATACATCACGGCAGACAAACAAAAGAATGGAATCAAAGCTAACTTCAAAATTCGCCACAACATTGAAGATGGATCCGTTCAACTAGCAGACCATTATCAACAAAATACTCCAATTGGCGATGGCCCTGTCCTTTTACCAGACAACCATTACCTGTCGACACAATCTGCCCTTTCGAAAGATCCCAACGAAAAGCGTGACCACATGGTCCTTCTTGAGTTTGTAACTGCTGCTGGGATTACACATGGCATGGATGAGCTCTACAAATAA"
-
-	sequenceLength := len(gfpSequence)
-
-	// Splitting the sequence into two parts to make a multi-location feature.
-	sequenceFirstHalf := gfpSequence[:sequenceLength/2]
-	sequenceSecondHalf := transform.ReverseComplement(gfpSequence[sequenceLength/2:]) // This feature is reverse complemented.
-
-	// rejoining the two halves into a single string where the second half of the sequence is reverse complemented.
-	gfpSequenceModified := sequenceFirstHalf + sequenceSecondHalf
-
-	// initialize sequence and feature structs.
-	var sequence genbank.Genbank
-	var feature genbank.Feature
-
-	// set the initialized sequence struct's sequence.
-	sequence.Sequence = gfpSequenceModified
-	// initialize sublocations to be usedin the feature.
-
-	var subLocation genbank.Location
-	var subLocationReverseComplemented genbank.Location
-
-	subLocation.Start = 0
-	subLocation.End = sequenceLength / 2
-
-	subLocationReverseComplemented.Start = sequenceLength / 2
-	subLocationReverseComplemented.End = sequenceLength
-	subLocationReverseComplemented.Complement = true // According to genbank complement means reverse complement. What a country.
-
-	feature.Description = "Green Flourescent Protein"
-	feature.Location.SubLocations = []genbank.Location{subLocation, subLocationReverseComplemented}
-
-	// Add the GFP feature to the sequence struct.
-	sequence.AddFeature(&feature)
-
-	// get the GFP feature sequence string from the sequence struct.
-	featureSequence, _ := feature.GetSequence()
-
-	// check to see if the feature was inserted properly into the sequence.
-	if gfpSequence != featureSequence {
-		t.Error("Feature sequence was not properly retrieved.")
-	}
-
-}
