+++ synthesis/codon/codon_test.go
@@ -78,7 +78,6 @@ func TestOptimize(t *testing.T) {
 	// weight our codon optimization table using the regions we collected from the genbank file above
 	optimizationTable := codonTable.OptimizeTable(codingRegions)
 
-func TestOptimize(t *testing.T) {
 	optimizedSequence, _ := Optimize(gfpTranslation, optimizationTable)
 	optimizedSequenceTranslation, _ := Translate(optimizedSequence, optimizationTable)
 
@@ -88,6 +87,25 @@ func TestOptimize(t *testing.T) {
 }
 
 func TestOptimizeSameSeed(t *testing.T) {
+	var gfpTranslation = "MASKGEELFTGVVPILVELDGDVNGHKFSVSGEGEGDATYGKLTLKFICTTGKLPVPWPTLVTTFSYGVQCFSRYPDHMKRHDFFKSAMPEGYVQERTISFKDDGNYKTRAEVKFEGDTLVNRIELKGIDFKEDGNILGHKLEYNYNSHNVYITADKQKNGIKANFKIRHNIEDGSVQLADHYQQNTPIGDGPVLLPDNHYLSTQSALSKDPNEKRDHMVLLEFVTAAGITHGMDELYK*"
+	var sequence, _ = genbank.Read("../../data/puc19.gbk")
+	var codonTable = GetCodonTable(11)
+
+	// a string builder to build a single concatenated string of all coding regions
+	var codingRegionsBuilder strings.Builder
+
+	// iterate through the features of the genbank file and if the feature is a coding region, append the sequence to the string builder
+	for _, feature := range sequence.Features {
+		if feature.Type == "CDS" {
+			sequence, _ := feature.GetSequence()
+			codingRegionsBuilder.WriteString(sequence)
+		}
+	}
+
+	// get the concatenated sequence string of the coding regions
+	codingRegions := codingRegionsBuilder.String()
+
+	var optimizationTable = codonTable.OptimizeTable(codingRegions)
 	randomSeed := 10
 
 	optimizedSequence, _ := Optimize(gfpTranslation, optimizationTable, randomSeed)
@@ -99,6 +117,26 @@ func TestOptimizeSameSeed(t *testing.T) {
 }
 
 func TestOptimizeDifferentSeed(t *testing.T) {
+	var gfpTranslation = "MASKGEELFTGVVPILVELDGDVNGHKFSVSGEGEGDATYGKLTLKFICTTGKLPVPWPTLVTTFSYGVQCFSRYPDHMKRHDFFKSAMPEGYVQERTISFKDDGNYKTRAEVKFEGDTLVNRIELKGIDFKEDGNILGHKLEYNYNSHNVYITADKQKNGIKANFKIRHNIEDGSVQLADHYQQNTPIGDGPVLLPDNHYLSTQSALSKDPNEKRDHMVLLEFVTAAGITHGMDELYK*"
+	var sequence, _ = genbank.Read("../../data/puc19.gbk")
+	var codonTable = GetCodonTable(11)
+
+	// a string builder to build a single concatenated string of all coding regions
+	var codingRegionsBuilder strings.Builder
+
+	// iterate through the features of the genbank file and if the feature is a coding region, append the sequence to the string builder
+	for _, feature := range sequence.Features {
+		if feature.Type == "CDS" {
+			sequence, _ := feature.GetSequence()
+			codingRegionsBuilder.WriteString(sequence)
+		}
+	}
+
+	// get the concatenated sequence string of the coding regions
+	codingRegions := codingRegionsBuilder.String()
+
+	var optimizationTable = codonTable.OptimizeTable(codingRegions)
+
 	optimizedSequence, _ := Optimize(gfpTranslation, optimizationTable)
 	otherOptimizedSequence, _ := Optimize(gfpTranslation, optimizationTable)
